==> notebooks/01_data_collection.ipynb <==
import requests
import json
import os
# URL da API
url = "https://guilhermeonrails.github.io/api-csharp-songs/songs.json"

# Requisição à API
response = requests.get(url)
songs_data = response.json()

# Visualizar uma amostra dos dados
print(songs_data[:5])


# Diretório para salvar os dados brutos
raw_data_dir = '../data/raw/'
os.makedirs(raw_data_dir, exist_ok=True)

# Salvar os dados brutos em um arquivo JSON
raw_data_path = os.path.join(raw_data_dir, 'songs.json')
with open(raw_data_path, 'w') as f:
    json.dump(songs_data, f)
