==> notebooks/03_data_split.ipynb <==
# Etapa 4: Divisão de Dados

import pandas as pd
from sklearn.model_selection import train_test_split

# Carregar os dados
data_path = '../data/processed/songs_processed.csv'
df = pd.read_csv(data_path)

# Dividir os dados em treinamento, validação e teste
train_data, test_data = train_test_split(df, test_size=0.2, random_state=42)
train_data, val_data = train_test_split(train_data, test_size=0.2, random_state=42)

# Salvar os dados divididos
train_data.to_csv('../data/splits/train_data.csv', index=False)
val_data.to_csv('../data/splits/val_data.csv', index=False)
test_data.to_csv('../data/splits/test_data.csv', index=False)

# Imprimir informações sobre os conjuntos de dados
print("Tamanho do conjunto de treinamento:", len(train_data))
print("Tamanho do conjunto de validação:", len(val_data))
print("Tamanho do conjunto de teste:", len(test_data))
